Extract pull-based FFI core
