+++ crates/ffi-core/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "minip2p-ffi-core"
+description = "Binding-agnostic embedding core for minip2p"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+publish = false
+
+[lints]
+workspace = true
+
+[dependencies]
+minip2p = { package = "minip2p-rs", path = "../minip2p", features = ["discovery", "mdns", "tcp"] }
+minip2p-swarm = { path = "../swarm" }
+minip2p-pubsub = { path = "../pubsub" }
+thiserror = "2.0.18"
+
+[dev-dependencies]
+minip2p-relay = { path = "../relay" }
+minip2p-test-support = { path = "../test-support" }
+serde_json = "1.0.149"
