+++ crates/ffi/Cargo.toml
@@ -16,14 +16,12 @@ crate-type = ["lib", "staticlib", "cdylib"]
 workspace = true
 
 [dependencies]
-minip2p = { package = "minip2p-rs", path = "../minip2p", features = ["discovery", "mdns", "tcp"] }
-minip2p-swarm = { path = "../swarm" }
-minip2p-pubsub = { path = "../pubsub" }
-thiserror = "2.0.18"
+minip2p-ffi-core = { path = "../ffi-core" }
 # Must remain paired with ubrn and @ubjs/core 0.31.0-3.
 uniffi = "=0.31.0"
 
 [dev-dependencies]
+minip2p = { package = "minip2p-rs", path = "../minip2p", features = ["discovery", "mdns", "tcp"] }
 minip2p-relay = { path = "../relay" }
 minip2p-test-support = { path = "../test-support" }
 serde_json = "1.0.149"
